requirements-dev.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
cloudinary==1.42.2
cryptography==45.0.4
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.9.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
oauthlib==3.3.1
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
